Added enemy script, gun scene and script
Created some basic visual mechanics of the player holding a gun that rotates around to point towards wherever the mouse is, using temporary assets (Made in 10 seconds in aseprite)

Also gave the enemy a script to make it walk towards the player. Nothing yet for it damaging the player or being killable.