Added projectiles that can be fired from the gun
Projectiles spawn at the barrel of the gun, and go in the direction of the mouse at the time they were shot (They always go straight no matter where the mouse moves after)

The code to make them come out of the correct place when aiming left is a bit jank, but works. That's Player.gd lines 57 and 61.